aoi-system/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.0.74
numpy==1.24.0
Pillow==10.0.0
werkzeug==3.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0

# PDF Generation
reportlab==4.0.7
python-docx==1.1.0

# Additional utilities
python-dotenv==1.0.0
